+++ fuzz/Cargo.toml
@@ -11,7 +11,6 @@ cargo-fuzz = true
 
 [dependencies]
 bytes = "1"
-lazy_static = "1"
 libfuzzer-sys = "0.4"
 
 [dependencies.retina]
