+++ fuzz/fuzz_targets/depacketize_jpeg.rs
@@ -1,17 +1,13 @@
 #![no_main]
-use lazy_static::lazy_static;
 use libfuzzer_sys::fuzz_target;
-use retina::Timestamp;
-use std::{num::NonZeroU32, sync::Mutex};
-
-lazy_static! {
-    static ref DEPACKETIZER: Mutex<retina::codec::Depacketizer> =
-        Mutex::new(retina::codec::Depacketizer::new("video", "jpeg", 90_000, None, None).unwrap());
-}
+use std::num::NonZeroU32;
 
 fuzz_target!(|data: &[u8]| {
     let mut data = data;
-
+    let mut depacketizer =
+        retina::codec::Depacketizer::new("video", "jpeg", 90_000, None, None).unwrap();
+    let mut timestamp = retina::Timestamp::new(0, NonZeroU32::new(90_000).unwrap(), 0).unwrap();
+    let mut sequence_number: u16 = 0;
     let conn_ctx = retina::ConnectionContext::dummy();
     let stream_ctx = retina::StreamContext::dummy();
     let pkt_ctx = retina::PacketContext::dummy();
@@ -20,6 +16,7 @@ fuzz_target!(|data: &[u8]| {
             Some(r) => r,
             None => return,
         };
+        let ts_change = (hdr & 0b001) != 0;
         let mark = (hdr & 0b010) != 0;
         let loss = (hdr & 0b100) != 0;
         let len = usize::from(hdr >> 3);
@@ -28,13 +25,18 @@ fuzz_target!(|data: &[u8]| {
         }
         let (payload, rest) = rest.split_at(len);
         data = rest;
-
+        if loss {
+            sequence_number = sequence_number.wrapping_add(1);
+        }
+        if ts_change {
+            timestamp = timestamp.try_add(1).unwrap();
+        }
         let pkt = retina::rtp::ReceivedPacketBuilder {
             ctx: pkt_ctx,
             stream_id: 0,
-            timestamp: Timestamp::new(0, NonZeroU32::new(96000).unwrap(), 0).unwrap(),
+            timestamp,
             ssrc: 0,
-            sequence_number: 1,
+            sequence_number,
             loss: u16::from(loss),
             payload_type: 96,
             mark,
@@ -42,18 +44,14 @@ fuzz_target!(|data: &[u8]| {
         .build(payload.iter().copied())
         .unwrap();
         //println!("pkt: {:#?}", pkt);
-        if DEPACKETIZER.lock().unwrap().push(pkt).is_err() {
+        if depacketizer.push(pkt).is_err() {
             return;
         }
-        while let Some(item) = DEPACKETIZER
-            .lock()
-            .unwrap()
-            .pull(&conn_ctx, &stream_ctx)
-            .transpose()
-        {
+        while let Some(item) = depacketizer.pull(&conn_ctx, &stream_ctx).transpose() {
             if item.is_err() {
                 return;
             }
         }
+        sequence_number = sequence_number.wrapping_add(1);
     }
 });
