--- fuzz_afl/.gitignore
@@ -1,3 +1,0 @@
-out/
-out-*/
-target/
